Fix syntactical issues and organized functions into separate folders
