+++ main.c
@@ -33,5 +33,7 @@ int main(void)
 	_printf("[%u]\n", 500);
 	_printf("Unsigned octal:[%o]\n", ui);
 	printf("Unsigned octal:[%o]\n", ui);
+	_printf("Unsigned hexadecimal:[%x, %X]\n", ui, ui);
+	printf("Unsigned hexadecimal:[%x, %X]\n", ui, ui);
 	return (0);
 }
