+++ printfunctions.c
@@ -41,6 +41,12 @@ int printstr(va_list list)
 	return (counter);
 }
 
+/**
+ *printint - prints an integer
+ *@list: variadic list of arguments
+ *
+ *Return: number of characters printed to stdout
+ */
 int printint(va_list list)
 {
 	int tens = 1;
@@ -72,125 +78,31 @@ int printint(va_list list)
 }
 
 /**
- *printbin - prints an unsigned int as binary
+ *printuint - prints an unsigned int
  *@list: variadic argument list
  *
  *Return: number of characters printed to stdout
  */
-int printbin(va_list list)
-{
-	int bin[32];
-	int i = 0;
-	unsigned int num = va_arg(list, unsigned int);
-	int counter = 0;
-
-	while (num != 0)
-	{
-		bin[i] = num % 2;
-		i++;
-		num /= 2;
-	}
-	for (i = i - 1; i >= 0; i--)
-	{
-		_putchar(bin[i] + '0');
-		counter++;
-	}
-	return (counter);
-}
-
 int printuint(va_list list)
 {
-        unsigned int tens = 1;
-        unsigned int num = va_arg(list, unsigned int);
-        unsigned int tensit = num;
-        unsigned int counter = 0;
-
-        while (num != 0)
-        {
-                num = num / 10;
-                tens *= 10;
-        }
-        tens /= 10;
-
-        while (tens != 0)
-        {
-                _putchar((tensit / tens) + '0');
-                tensit = tensit % tens;
-                tens /= 10;
-                counter++;
-        }
-        return (counter);
-}
-
-int printoctal(va_list list)
-{
-	int i = 0;
-	unsigned int oct[32];
-	unsigned int counter;
-	unsigned int num = va_arg(list, unsigned);
-
-	while (num != 0)
-	{
-		oct[i] = num % 8;
-		num /= 8;
-		i++;
-	}
-	for (i = i - 1; i >= 0; i--)
-	{
-		_putchar(oct[i] + '0');
-		counter++;
-	}
-	return (counter);
-}
-
-int printhex(va_list list)
-{
-	int i = 0;
-	unsigned int hex[32];
-	char letters[] = {'a', 'b', 'c', 'd', 'e', 'f'};
+	unsigned int tens = 1;
+	unsigned int num = va_arg(list, unsigned int);
+	unsigned int tensit = num;
 	unsigned int counter = 0;
-	unsigned int num = va_arg(list, unsigned);
 
 	while (num != 0)
 	{
-		if ((num % 16) > 9)
-			hex[i] = letters[(num % 16) % 10];
-		else
-			hex[i] = (num % 16) + '0';
-		i++;
-		num = num % 16;
+		num = num / 10;
+		tens *= 10;
 	}
+	tens /= 10;
 
-	for (i = i - 1; i >= 0; i--)
+	while (tens != 0)
 	{
-		_putchar(hex[i]);
+		_putchar((tensit / tens) + '0');
+		tensit = tensit % tens;
+		tens /= 10;
 		counter++;
 	}
 	return (counter);
-}
-
-int printhexcaps(va_list list)
-{
-        int i = 0;
-        unsigned int hex[32];
-        char letters[] = {'A', 'B', 'C', 'D', 'E', 'F'};
-        unsigned int counter = 0;
-        unsigned int num = va_arg(list, unsigned);
-
-        while (num != 0)
-        {
-                if ((num % 16) > 9)
-                        hex[i] = letters[(num % 16) % 10];
-                else
-                        hex[i] = (num % 16) + '0';
-                i++;
-                num = num % 16;
-        }
-
-        for (i = i - 1; i >= 0; i--)
-        {
-                _putchar(hex[i]);
-                counter++;
-        }
-        return (counter);
 }
